feat: add sync script

+++ backend/tests/integration/setup.ts
@@ -7,7 +7,9 @@ before(async () => {
   sinon.stub(PostmarkClient, 'sendEmail').resolves();
 
   await sequelize.authenticate();
-  await sequelize.sync({ force: true });
+  if (process.env.NODE_ENV === 'test') {
+    await sequelize.sync({ force: true });
+  }
 });
 
 after(async () => {
